sparse vector_ordering: deal with vector indices as 'size_t'
This pays down some cast debt from previous changes, while moving
towards the goal of pervasive `size_t` indexing. This change causes some
flow on effects in calling code which reduce debt.

+++ cmd/gvmap/make_map.c
@@ -95,7 +95,7 @@ void map_palette_optimal_coloring(char *color_scheme, SparseMatrix A0,
 }
 
 void map_optimal_coloring(int seed, SparseMatrix A, float *rgb_r,  float *rgb_g, float *rgb_b){
-  int *p = NULL;
+  size_t *p = NULL;
   float *u = NULL;
   const size_t n = A->m;
